+++ pyluos/modules/__init__.py
@@ -11,6 +11,9 @@ from .potard import Potard
 from .servo import Servo
 from .relay import Relay
 from .led import Led
+from .eddy import Eddy
+from .handy import Handy
+
 
 __all__ = [
     'name2mod',
@@ -27,6 +30,8 @@ __all__ = [
     'Relay',
     'Led',
     'L0DCMotor',
+    'Eddy',
+    'Handy',
 ]
 
 name2mod = {
@@ -43,4 +48,6 @@ name2mod = {
     'servo': Servo,
     'relay': Relay,
     'led': Led,
+    'eddy': Eddy,
+    'handy': Handy
 }